+++ addition_bs.html
@@ -49,7 +49,7 @@
               <button id="secondNumGen" class="btn btn-outline-primary btn-sm w-10" disabled></button>
               <button id="equalTo" class="btn btn-sm w-10" disabled>=</button>
               <input type="text" id="answer" class="w-10" value="" onkeypress="return isNumber(event)" />&nbsp;&nbsp;
-              <input type="button" id="submit_answer" class="btn btn-outline-primary btn-sm w-10" onclick="scoreMark()"
+              <input type="button" id="submitAnswer" class="btn btn-outline-primary btn-sm w-10" onclick="scoreMark()"
                 value="GO" />
             </div>
             <div id="studentPracticeSummary" class="form-group row">
@@ -62,7 +62,7 @@
             <table id="practicedResults" border="1">
               <tr>
                 <th colspan="6">Result Table<br />
-                  <div id="date_time">
+                  <div id="dateTime">
 
                   </div>
                 </th>
