+++ scripts/model/AnswerTips.js
@@ -1,0 +1,27 @@
+function explainMultiplication([a, b, ...rest]) {
+  const tens = Math.floor(b / 10) * 10;
+  const ones = Math.floor(b % 10);
+  const tensMultiples = tens * a;
+  const onesMultiples = ones * a;
+  const tensString = `${tens} x ${a} = ${tens * a}`;
+  const onesString = `${ones} x ${a} = ${ones * a}`;
+  const total = `${tensMultiples} + ${onesMultiples} = ${
+    tensMultiples + onesMultiples}`;
+  return ones != 0
+    ? `${onesString}<br>${tensString}<br>${total}`
+    : `${tensString}<br>${total}`;
+}
+
+export default function explanation(mathOperations, inputs) {
+  const numbers = [...inputs];
+  numbers.sort((a, b) => a - b);
+  if (mathOperations === 'multiplication') {
+    if (numbers[1] > 10 && numbers[0] < 10) {
+      return explainMultiplication(numbers);
+    } if (numbers[1] > 10 && numbers[0] > 10) {
+      return explainMultiplication([numbers[1], numbers[0]]);
+    }
+  }
+  return '';
+}
+
