+++ scripts/model/Question.js
@@ -1,0 +1,9 @@
+export default class  Question {
+  constructor(firstNum, secondNum, operation, submittedAnswer, submissionTime) {
+    this.firstNum = firstNum;
+    this.secondNum = secondNum;
+    this.operation = operation;
+    this.submittedAnswer = submittedAnswer;
+    this.submissionTime = submissionTime;
+  }
+}
