+++ scripts/package.json
@@ -11,6 +11,8 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@babel/preset-env": "7.10.4",
+    "babel-jest": "26.1.0",
     "eslint": "7.4.0",
     "eslint-config-airbnb": "18.2.0",
     "eslint-plugin-import": "2.22.0",
