+++ scripts/random.js
@@ -1,6 +1,4 @@
-class Random {
-  constructor() {}
-
+export default class Random {
   /**
    * Credit : https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/random
    * @param {*} min
@@ -14,9 +12,6 @@ class Random {
 
   static getRandomIntInclusiveWithExceptions(min, max, ...excludes) {
     const result = this.getRandomIntInclusive(min, max);
-    console.log(`intermediate result ${result}`);
     return excludes.indexOf(result) == -1 ? result : this.getRandomIntInclusiveWithExceptions(min, max, excludes);
   }
 }
-
-module.exports = { Random };
