+++ subtraction.html
@@ -26,7 +26,7 @@
           <form id="formPractice">
             <b><label id="welcomeMessage" /></b>&nbsp;&nbsp;
             <input type="hidden" id="operations" value="subtraction" />
-            <input type="hidden" id="max_input" value="40" />
+            <input type="hidden" id="maxInput" value="40" />
 
             <div id="level" style="visibility: hidden;">
               <label for="SquareMode">Square mode</label>&nbsp;&nbsp;&nbsp;
@@ -38,7 +38,7 @@
             <input type="text" id="secondNumGen" value="" readonly="readonly" /><br />
             <label for="answer">Answer</label>&nbsp;&nbsp;&nbsp;
             <input type="text" id="answer" value="" onkeypress="return true;" />
-            <label for="submit_answer"><input type="button" id="submit_answer" onclick="scoreMark()"
+            <label for="submitAnswer"><input type="button" id="submitAnswer" onclick="scoreMark()"
                 value="Submit Answer" /></label>&nbsp;&nbsp;&nbsp;
 
             <br /><br />
@@ -59,7 +59,7 @@
             <table id="practicedResults" border="1">
               <tr>
                 <th colspan="6">Result Table<br />
-                  <div id="date_time">
+                  <div id="dateTime">
 
                   </div>
                 </th>
